Don't overwrite ephemeral_devices or format ephemeral devices if they have the correct filesystem already

+++ recipes/default.rb
@@ -43,6 +43,7 @@ else
 
       execute "Reformat Ephemeral Device: #{dev[:device]}" do
         command "mkfs.#{node['ephemeral_mdadm']['filesystem']} #{dev[:device]}"
+        not_if { node['ephemeral_mdadm']['filesystem'] == dev[:fs_type] }
       end
 
     end
